requirements.txt
openai==1.3.8
python-telegram-bot==20.7
APScheduler==3.10.4
pytz==2023.3
python-dotenv==1.0.0
nest-asyncio==1.5.8